[Cleanup] Replaced all tabs with 4 spaces (which is the project's default whenever you hit a tab anyway)

+++ Runtime/Logging/Log.cpp
@@ -33,17 +33,17 @@ using namespace Spartan::Math;
 
 namespace Spartan
 {
-	weak_ptr<ILogger> Log::m_logger;
-	ofstream Log::m_fout;
-	mutex Log::m_mutex_log;
+    weak_ptr<ILogger> Log::m_logger;
+    ofstream Log::m_fout;
+    mutex Log::m_mutex_log;
     vector<LogCmd> Log::m_log_buffer;
-	string Log::m_log_file_name	    = "log.txt";
-	bool Log::m_log_to_file		    = true; // start logging to file (unless changed by the user, e.g. Renderer initialization was successful, so logging can happen on screen)
-	bool Log::m_first_log		    = true;
+    string Log::m_log_file_name        = "log.txt";
+    bool Log::m_log_to_file            = true; // start logging to file (unless changed by the user, e.g. Renderer initialization was successful, so logging can happen on screen)
+    bool Log::m_first_log            = true;
    
-	// Everything resolves to this
-	void Log::Write(const char* text, const LogType type)
-	{
+    // Everything resolves to this
+    void Log::Write(const char* text, const LogType type)
+    {
         if (!text)
         {
             LOG_ERROR_INVALID_PARAMETER();
@@ -64,40 +64,40 @@ namespace Spartan
             FlushBuffer();
             LogString(text, type);
         }
-	}
+    }
 
     void Log::WriteFInfo(const char* text, ...)
-	{
-		char buffer[1024];
-		va_list args;
-		va_start(args, text);
-		auto w = vsnprintf(buffer, sizeof(buffer), text, args);
-		va_end(args);
+    {
+        char buffer[1024];
+        va_list args;
+        va_start(args, text);
+        auto w = vsnprintf(buffer, sizeof(buffer), text, args);
+        va_end(args);
 
-		Write(buffer, LogType::Info);
-	}
+        Write(buffer, LogType::Info);
+    }
 
     void Log::WriteFWarning(const char* text, ...)
-	{
-		char buffer[1024];
-		va_list args;
-		va_start(args, text);
-		auto w = vsnprintf(buffer, sizeof(buffer), text, args);
-		va_end(args);
+    {
+        char buffer[1024];
+        va_list args;
+        va_start(args, text);
+        auto w = vsnprintf(buffer, sizeof(buffer), text, args);
+        va_end(args);
 
-		Write(buffer, LogType::Warning);
-	}
+        Write(buffer, LogType::Warning);
+    }
 
     void Log::WriteFError(const char* text, ...)
-	{
-		char buffer[1024];
-		va_list args;
-		va_start(args, text);
-		auto w = vsnprintf(buffer, sizeof(buffer), text, args);
-		va_end(args);
+    {
+        char buffer[1024];
+        va_list args;
+        va_start(args, text);
+        auto w = vsnprintf(buffer, sizeof(buffer), text, args);
+        va_end(args);
 
-		Write(buffer, LogType::Error);
-	}
+        Write(buffer, LogType::Error);
+    }
 
     void Log::Write(const string& text, const LogType type)
     {
@@ -138,39 +138,39 @@ namespace Spartan
     }
 
     void Log::Write(const weak_ptr<Entity>& entity, const LogType type)
-	{
-		entity.expired() ? Write("Null", type) : Write(entity.lock()->GetName(), type);
-	}
+    {
+        entity.expired() ? Write("Null", type) : Write(entity.lock()->GetName(), type);
+    }
 
-	void Log::Write(const std::shared_ptr<Entity>& entity, const LogType type)
-	{
-		entity ? Write(entity->GetName(), type) : Write("Null", type);
-	}
+    void Log::Write(const std::shared_ptr<Entity>& entity, const LogType type)
+    {
+        entity ? Write(entity->GetName(), type) : Write("Null", type);
+    }
 
-	void Log::Write(const Vector2& value, const LogType type)
-	{
-		Write(value.ToString(), type);
-	}
+    void Log::Write(const Vector2& value, const LogType type)
+    {
+        Write(value.ToString(), type);
+    }
 
-	void Log::Write(const Vector3& value, const LogType type)
-	{
-		Write(value.ToString(), type);
-	}
+    void Log::Write(const Vector3& value, const LogType type)
+    {
+        Write(value.ToString(), type);
+    }
 
-	void Log::Write(const Vector4& value, const LogType type)
-	{
-		Write(value.ToString(), type);
-	}
+    void Log::Write(const Vector4& value, const LogType type)
+    {
+        Write(value.ToString(), type);
+    }
 
-	void Log::Write(const Quaternion& value, const LogType type)
-	{
-		Write(value.ToString(), type);
-	}
+    void Log::Write(const Quaternion& value, const LogType type)
+    {
+        Write(value.ToString(), type);
+    }
 
-	void Log::Write(const Matrix& value, const LogType type)
-	{
-		Write(value.ToString(), type);
-	}
+    void Log::Write(const Matrix& value, const LogType type)
+    {
+        Write(value.ToString(), type);
+    }
 
     void Log::FlushBuffer()
     {
@@ -186,17 +186,6 @@ namespace Spartan
     }
 
     void Log::LogString(const char* text, const LogType type)
-	{
-        if (!text)
-        {
-            LOG_ERROR_INVALID_PARAMETER();
-            return;
-        }
-
-		m_logger.lock()->Log(string(text), static_cast<uint32_t>(type));
-	}
-
-	void Log::LogToFile(const char* text, const LogType type)
     {
         if (!text)
         {
@@ -204,26 +193,37 @@ namespace Spartan
             return;
         }
 
-		const string prefix		= (type == LogType::Info) ? "Info:" : (type == LogType::Warning) ? "Warning:" : "Error:";
-		const auto final_text	= prefix + " " + text;
+        m_logger.lock()->Log(string(text), static_cast<uint32_t>(type));
+    }
 
-		// Delete the previous log file (if it exists)
-		if (m_first_log)
-		{
-			FileSystem::Delete(m_log_file_name);
-			m_first_log = false;
-		}
+    void Log::LogToFile(const char* text, const LogType type)
+    {
+        if (!text)
+        {
+            LOG_ERROR_INVALID_PARAMETER();
+            return;
+        }
 
-		// Open/Create a log file to write the error message to
-		m_fout.open(m_log_file_name, ofstream::out | ofstream::app);
+        const string prefix        = (type == LogType::Info) ? "Info:" : (type == LogType::Warning) ? "Warning:" : "Error:";
+        const auto final_text    = prefix + " " + text;
 
-		if (m_fout.is_open())
-		{
-			// Write out the error message
-			m_fout << final_text << endl;
+        // Delete the previous log file (if it exists)
+        if (m_first_log)
+        {
+            FileSystem::Delete(m_log_file_name);
+            m_first_log = false;
+        }
 
-			// Close the file
-			m_fout.close();
-		}
-	}
+        // Open/Create a log file to write the error message to
+        m_fout.open(m_log_file_name, ofstream::out | ofstream::app);
+
+        if (m_fout.is_open())
+        {
+            // Write out the error message
+            m_fout << final_text << endl;
+
+            // Close the file
+            m_fout.close();
+        }
+    }
 }
